requirements.txt
waveshare-epaper
python-dotenv
requests
